Fix mac x64 build error in pepper_truetype_font_list_mac.mm.
Introduced http://crrev.com/195082

BUG=230130
TBR=bbudge@chromium.org, raymes@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14392002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@195465 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4a2e8d5f38da690160909cdbde30010ae6c1e4ed

+++ content/browser/renderer_host/pepper/pepper_truetype_font_list_mac.mm
@@ -56,7 +56,7 @@ void GetFontsInFamily_SlowBlocking(
     ppapi::proxy::SerializedTrueTypeFontDesc desc;
     desc.family = family;
     NSInteger font_weight = [[font_info objectAtIndex:2] intValue];
-    font_weight = std::max(0, font_weight);
+    font_weight = std::max(static_cast<NSInteger>(0), font_weight);
     font_weight = std::min(kPepperFontWeightsLength - 1, font_weight);
     desc.weight = kPepperFontWeights[font_weight];
